requirements.txt
requests==2.31.0
PyJWT==2.10.1
cryptography==44.0.0

